Update cast-and-convert-transact-sql.md
added a section to the table to call out the rounding rules when casting a datetime to a time.. The rounding behavior seems to me to be undocumented or a bug. It definitely was unexpected see this post on SO for further discussion: https://stackoverflow.com/questions/69033081/t-sql-truncate-time-not-round-to-the-nearest-minute it's proably worth adding an example to the PR to demonstrate the behavior. If I can get confirmation that this behavior is by design and not a bug from the product team I will be glad to amend the PR.